Move the TextureShaderApplier out to DepalettizeCommon
